pyproject.toml
[project]
name = "mtg-collector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "requests>=2.32.5",
]
